--- Arrays/Q11. Rotate Image/Optimized-Approach.cpp
@@ -1,36 +1,0 @@
-#include<bits/stdc++.h>
-using namespace std;
-
-void rotate(vector<vector<int>>& matrix) {
-    int n = matrix.size();
-
-    // Step 1: Transpose the matrix (swap rows with columns)
-    for (int i = 0; i < n; i++) {
-        for (int j = 0; j < i; j++) {
-            swap(matrix[i][j], matrix[j][i]); // Swap elements across the diagonal
-        }
-    }
-
-    // Step 2: Reverse each row
-    for (int i = 0; i < n; i++) {
-        reverse(matrix[i].begin(), matrix[i].end()); // Reverse the row
-    }
-}
-
-int main() {
-    vector<vector<int>> arr = {{1, 2, 3}, {4, 5, 6}, {7, 8, 9}};
-    
-    // Perform rotation in-place
-    rotate(arr);
-    
-    // Output the rotated matrix
-    cout << "Rotated Image:" << endl;
-    for (int i = 0; i < arr.size(); i++) {
-        for (int j = 0; j < arr[0].size(); j++) {
-            cout << arr[i][j] << " ";
-        }
-        cout << endl;
-    }
-
-    return 0;
-}
