--- Linked List/Q4. Merge In Between Linked Lists/OptimalApproach.txt
@@ -1,35 +1,0 @@
-
-# Optimized Approach: Detailed Explanation
-
-### Problem Intuition:
-The problem involves merging `list2` into `list1` by removing the nodes in `list1` from index `a` to index `b` (inclusive), and inserting the entirety of `list2` in place of these removed nodes. The objective is to minimize the number of traversals and efficiently merge the lists in a single pass.
-
-### Approach:
-1. **Finding the Key Nodes:**
-   - **`pre` Node (Node before the removal segment in `list1`):** 
-     - Traverse `list1` to find the node at index `a-1`. This is the node just before the start of the removal range (index `a`).
-   - **`post` Node (Node after the removal segment in `list1`):**
-     - Traverse `list1` to find the node at index `b+1`. This is the node just after the end of the removal range (index `b`).
-   - **Tail of `list2`:**
-     - Traverse `list2` to find its last node. This will help in connecting the end of `list2` to the `post` node of `list1`.
-
-2. **Modifying the Links:**
-   - **Connecting `list1` to `list2`:**
-     - Set the `next` pointer of the `pre` node (found at index `a-1`) to the head of `list2`. This effectively links the portion of `list1` before index `a` to the start of `list2`.
-   - **Connecting the Tail of `list2` to `list1`:**
-     - After `list2` is linked, find the tail of `list2` (the last node) and set its `next` pointer to the `post` node (found at index `b+1`), which links the end of `list2` to the remaining part of `list1`.
-
-3. **Returning the Modified List:**
-   - After the connections are made, the modified `list1` will contain the merged nodes, with the segment from index `a` to `b` replaced by `list2`. Return the updated `list1` as the result.
-
-### Time Complexity:
-- **Time Complexity:** O(N), where `N` is the number of nodes in `list1`. We traverse `list1` once to find the `pre` and `post` nodes, and we traverse `list2` once to find its tail. Thus, we only perform linear traversals over both `list1` and `list2`.
-  
-- **Space Complexity:** O(1), because we are only manipulating pointers, and no extra data structures are created. The space usage remains constant.
-
-### Explanation of Time Complexity:
-- The time complexity is dominated by the fact that we need to traverse both `list1` and `list2` to find key nodes and modify the links. Since there are no nested loops or additional space-consuming operations, the total time complexity is linear relative to the size of `list1` and `list2`, which is O(N), where `N` is the size of the larger list.
-
-### Final Thoughts:
-- This approach is optimal in terms of both time and space complexity. The main advantage is that we minimize the number of traversals required (only two for `list1` and one for `list2`), making it efficient for large lists.
-- The approach ensures that the linked lists are merged correctly with minimal overhead, making it suitable for handling lists with sizes up to the constraint limits (10^4 nodes).
